Add setting helpers.
Setting helpers allow a setting to specify an object implementing ISettingHelper to help the user enter a value. When the helper is not null, whatever control the setting would normally show to let the user input a value is hidden and instead a button is shown with the text specified by ISettingHelper.HelpButtonText.