Finished web behaviours
Web behaviours will now be executed when a page, section or control is refreshed.
Added BackgroundTask class to give a background task way of letting the studio know it's important and why.
Changed the shutdown implementation so the studio checks for running background tasks and warns the user if there are any important ones.
Added SettingsChanges class to communicate changes in settings.
Added static GetChanges() method to SettingsCollection to get a SettingsChanges object.
Added packaging system to package extensions. (and any other files)
Added RegisterPackageDefinition() method to the IStudioEnvironment.