Complete UI overhaul.
Added tabs and a StudioView class to represent the class content.
Added a lot more events and methods to the `IStudioEnvironment` class.
Changed all windows to be responsive.
Updated README with a concrete feature set for V1.0
updated version to 0.2-beta.
Added home tab.
Added spinner for indeterminate background tasks to preserve space in the statusbar.
Remove unused namespaces project wide.
Changed menubar style so it blends with the default theme of the app.